// src/src/models/Project/IProjectSetting.tsx
export interface projectSettingsInterface {

    id: number;
    startDate: string;
    endDate: string;
    depositMoney: number;
    departM: string;
    firstN: string;
    lastN: string;
    nickN: string;
    position: string;
    secT: string;
    timeEnd: string;
    department_name: string;
    email: string;
    mobile : string;
}

export interface detailmanagerInterface {

    key: string;
    name: string;
    nametitle_t: string;
    firstname_t: string;
    lastname_t: string;
    department: string;
    email: string;
    role_name_t: string;
}